+++ flink-runtime/src/test/java/org/apache/flink/runtime/metrics/groups/JobManagerGroupTest.java
@@ -22,8 +22,8 @@ import org.apache.flink.api.common.JobID;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.MetricOptions;
 import org.apache.flink.runtime.jobgraph.JobGraphBuilder;
-import org.apache.flink.runtime.metrics.MetricRegistryConfiguration;
 import org.apache.flink.runtime.metrics.MetricRegistryImpl;
+import org.apache.flink.runtime.metrics.MetricRegistryTestUtils;
 import org.apache.flink.runtime.metrics.dump.QueryScopeInfo;
 import org.apache.flink.runtime.metrics.util.DummyCharacterFilter;
 import org.apache.flink.util.TestLogger;
@@ -45,7 +45,7 @@ public class JobManagerGroupTest extends TestLogger {
     public void addAndRemoveJobs() throws Exception {
         MetricRegistryImpl registry =
                 new MetricRegistryImpl(
-                        MetricRegistryConfiguration.defaultMetricRegistryConfiguration());
+                        MetricRegistryTestUtils.defaultMetricRegistryConfiguration());
         final JobManagerMetricGroup group = new JobManagerMetricGroup(registry, "localhost");
 
         final JobID jid1 = new JobID();
@@ -94,7 +94,7 @@ public class JobManagerGroupTest extends TestLogger {
     public void testCloseClosesAll() throws Exception {
         MetricRegistryImpl registry =
                 new MetricRegistryImpl(
-                        MetricRegistryConfiguration.defaultMetricRegistryConfiguration());
+                        MetricRegistryTestUtils.defaultMetricRegistryConfiguration());
         final JobManagerMetricGroup group = new JobManagerMetricGroup(registry, "localhost");
 
         final JobID jid1 = new JobID();
@@ -132,7 +132,7 @@ public class JobManagerGroupTest extends TestLogger {
     public void testGenerateScopeDefault() throws Exception {
         MetricRegistryImpl registry =
                 new MetricRegistryImpl(
-                        MetricRegistryConfiguration.defaultMetricRegistryConfiguration());
+                        MetricRegistryTestUtils.defaultMetricRegistryConfiguration());
         JobManagerMetricGroup group = new JobManagerMetricGroup(registry, "localhost");
 
         assertArrayEquals(new String[] {"localhost", "jobmanager"}, group.getScopeComponents());
@@ -146,7 +146,7 @@ public class JobManagerGroupTest extends TestLogger {
         Configuration cfg = new Configuration();
         cfg.setString(MetricOptions.SCOPE_NAMING_JM, "constant.<host>.foo.<host>");
         MetricRegistryImpl registry =
-                new MetricRegistryImpl(MetricRegistryConfiguration.fromConfiguration(cfg));
+                new MetricRegistryImpl(MetricRegistryTestUtils.fromConfiguration(cfg));
 
         JobManagerMetricGroup group = new JobManagerMetricGroup(registry, "host");
 
@@ -161,7 +161,7 @@ public class JobManagerGroupTest extends TestLogger {
     public void testCreateQueryServiceMetricInfo() {
         MetricRegistryImpl registry =
                 new MetricRegistryImpl(
-                        MetricRegistryConfiguration.defaultMetricRegistryConfiguration());
+                        MetricRegistryTestUtils.defaultMetricRegistryConfiguration());
         JobManagerMetricGroup jm = new JobManagerMetricGroup(registry, "host");
 
         QueryScopeInfo.JobManagerQueryScopeInfo info =
